phoneticMatch-component
